Renames auditor package.
