+++ core/src/main/java/gyro/core/command/AbstractCommand.java
@@ -22,8 +22,8 @@ import java.util.Optional;
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import gyro.core.GyroCore;
-import gyro.core.auditor.GyroAuditor;
-import gyro.core.metadata.MetadataDirectiveProcessor;
+import gyro.core.audit.GyroAuditor;
+import gyro.core.audit.MetadataDirectiveProcessor;
 import io.airlift.airline.Option;
 import io.airlift.airline.OptionType;
 import org.slf4j.LoggerFactory;
