+++ core/src/main/java/gyro/core/scope/RootScope.java
@@ -37,8 +37,9 @@ import gyro.core.GyroInputStream;
 import gyro.core.GyroOutputStream;
 import gyro.core.LogDirectiveProcessor;
 import gyro.core.PrintDirectiveProcessor;
-import gyro.core.auditor.AuditorDirectiveProcessor;
-import gyro.core.auditor.AuditorPlugin;
+import gyro.core.audit.AuditorDirectiveProcessor;
+import gyro.core.audit.AuditorPlugin;
+import gyro.core.audit.MetadataDirectiveProcessor;
 import gyro.core.auth.CredentialsDirectiveProcessor;
 import gyro.core.auth.CredentialsPlugin;
 import gyro.core.auth.UsesCredentialsDirectiveProcessor;
@@ -54,7 +55,6 @@ import gyro.core.diff.GlobalChangePlugin;
 import gyro.core.directive.DirectivePlugin;
 import gyro.core.directive.DirectiveSettings;
 import gyro.core.finder.FinderPlugin;
-import gyro.core.metadata.MetadataDirectiveProcessor;
 import gyro.core.plugin.PluginDirectiveProcessor;
 import gyro.core.plugin.PluginSettings;
 import gyro.core.reference.FinderReferenceResolver;
